pm_list_data: Rename private_message to direct_message.
We want to rename `pm` or `private_message` to `direct_message`
over the long term and part of it is to not use the old naming for
new code. This commit is a preparatory commit for introducing direct
messages filter.
We have not changed the file name yet, since that would be better done
for similar files in the same PR, we've just renamed the variables for
the file.

Part of the private_message to direct_message rename project.